Feature: Add decryption capabilities to psptool
This commit adds a new flag ("-c / --decrypt") that tries to decrypt
entires using unwrapped IKEK's. If possible, psptool tries to verify the
integrity if an entry after it is decrypted. If that succeeds, the
decryption worked. If not, there is no good way to verify the
decryption. Manual validation of the result is required.