psptool/{file.py,key_store_file.py}: Improve parsing of entries
Improve parsing of the entries in the directories. Sometimes
key stores may also be compressed, so it is necessary to extend
the basic file class with compression handling. Use the compression
library check for compression and decompress the key store entries
if neeeded.

Also extend the list of entries without a header, so that the parsing
does not fail.

Signed-off-by: Michał Żygowski <michal.zygowski@3mdeb.com>