psptool: Workaround entries crossing 16MB boundaries
Some images do not repsect the reqirement of placing the components
on 16MB flash page boundary. Typically those are 32MB images with
single ROMs. Treat the ROM on the first page as if it was a ROM
spanning whole image size to avoid file offsets getting out of bounds
of parent buffers.

Also fixup the physical addresses if the ROM size is bigger than 16MB.

Signed-off-by: Michał Żygowski <michal.zygowski@3mdeb.com>